Logger deleted. Useless comments deleted. Connection class path updated as it has been moved to connection package.
