+++ src/main/java/org/netcracker/educationcenter/elasticsearch/search/FTPFileObjectSearch.java
@@ -2,7 +2,7 @@ package org.netcracker.educationcenter.elasticsearch.search;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.netcracker.educationcenter.elasticsearch.Connection;
+import org.netcracker.educationcenter.elasticsearch.connection.Connection;
 import org.netcracker.educationcenter.elasticsearch.database.DatabaseConstants;
 
 /**
@@ -26,14 +26,6 @@ public class FTPFileObjectSearch implements DocumentSearch {
      */
     public FTPFileObjectSearch(Connection connection) {
         this.connection = connection;
-    }
-
-    /**
-     * @return current logger instance
-     */
-    @Override
-    public Logger getLogger() {
-        return LOG;
     }
 
     /**
